Leaderboard Table Component created
Component that has the leaderboard data table.
